odontocare/citas-service/requirements.txt
flask
flask_sqlalchemy
PyJWT
python-dotenv
jsonschema
requests